+++ api/index.py
@@ -1,3 +1,3 @@
-from app.main import app
+from sis_facturador.main import app
 
 __all__ = ["app"]